move mock check_sudo_check logic to new bootstrapper test file
