+++ src/core/tests/Test_CoreMain.py
@@ -1215,79 +1215,6 @@ class TestCoreMain(unittest.TestCase):
         os.remove = self.backup_os_remove
         runtime.stop()
 
-    def test_check_sudo_status_all_attempts_failed(self):
-        # Set raise_if_not_sudo=False to test the `return False` all attempts failed
-        argument_composer = ArgumentComposer()
-        argument_composer.operation = Constants.ASSESSMENT
-        runtime = RuntimeCompositor(argument_composer.get_composed_arguments(), legacy_mode=True, package_manager_name=Constants.APT,
-            set_mock_sudo_status='Always_False')
-
-        result = runtime.bootstrapper.check_sudo_status(raise_if_not_sudo=False)
-        self.assertFalse(result, "Expected check_sudo_status to return False after all attempts failed")
-        runtime.stop()
-
-    def test_check_sudo_status_throw_exception(self):
-        # Set raise_if_not_sudo=True to throw exception
-        argument_composer = ArgumentComposer()
-        argument_composer.operation = Constants.ASSESSMENT
-        runtime = RuntimeCompositor(argument_composer.get_composed_arguments(), legacy_mode=True, package_manager_name=Constants.APT,
-            set_mock_sudo_status='Always_False')
-
-        # Run check_sudo_status and expect an exception after all retries
-        with self.assertRaises(Exception) as context:
-            runtime.bootstrapper.check_sudo_status(raise_if_not_sudo=True)
-
-        # Verify exception msg contains the expected failure text
-        self.assertTrue("Unable to invoke sudo successfully" in str(context.exception))
-        runtime.stop()
-
-    def test_check_sudo_status_insufficient_output_lines(self):
-        # Test insufficient output lines to raise exception
-        argument_composer = ArgumentComposer()
-        argument_composer.operation = Constants.ASSESSMENT
-        runtime = RuntimeCompositor(argument_composer.get_composed_arguments(), legacy_mode=True, package_manager_name=Constants.APT,
-            set_mock_sudo_status='insufficient_output_lines')
-
-        # Run check_sudo_status and expect an exception after all retries
-        with self.assertRaises(Exception) as context:
-            runtime.bootstrapper.check_sudo_status()
-
-        # Verify exception msg contains the expected failure text
-        self.assertTrue("Unexpected sudo check result" in str(context.exception))
-        runtime.stop()
-
-    def test_check_sudo_status_unexpected_output_lines(self):
-        # Test unexpected output neither false or true to raise exception
-        argument_composer = ArgumentComposer()
-        argument_composer.operation = Constants.ASSESSMENT
-        runtime = RuntimeCompositor(argument_composer.get_composed_arguments(), legacy_mode=True, package_manager_name=Constants.APT,
-            set_mock_sudo_status='unexpected_output')
-
-        # Run check_sudo_status and expect an exception after all retries
-        with self.assertRaises(Exception) as context:
-            runtime.bootstrapper.check_sudo_status()
-
-        # Verify exception msg contains the expected failure text
-        self.assertTrue("Unexpected sudo check result" in str(context.exception))
-        runtime.stop()
-
-    def test_check_sudo_status_succeeds_on_third_attempt(self):
-        # Test retry logic in check sudo status after 2 failed attempts followed by success (true)
-        argument_composer = ArgumentComposer()
-        argument_composer.operation = Constants.ASSESSMENT
-        runtime = RuntimeCompositor(argument_composer.get_composed_arguments(), legacy_mode=True, package_manager_name=Constants.APT,
-            set_mock_sudo_status='Retry_True')
-
-        # Attempt to check sudo status, succeed (true) on the 3rd attempt
-        result = runtime.bootstrapper.check_sudo_status(raise_if_not_sudo=True)
-
-        # Verify the result is success (True)
-        self.assertTrue(result, "Expected check_sudo_status to succeed on the 3rd attempts")
-
-        # Verify 3 attempts were made
-        self.assertEqual(runtime.sudo_check_status_attempts, 3, "Expected exactly 4 attempts in check_sudo_status")
-        runtime.stop()
-
     def __check_telemetry_events(self, runtime):
         all_events = os.listdir(runtime.telemetry_writer.events_folder_path)
         self.assertTrue(len(all_events) > 0)
